Fix building with INTERN_DISABLE_ONNX
