+++ caffe2/utils/CMakeLists.txt
@@ -11,7 +11,7 @@ if(NOT BUILD_CAFFE2 OR INTERN_BUILD_MOBILE)
     )
   endif()
 
-  if(NOT BUILD_CAFFE2 AND NOT INTERN_BUILD_MOBILE)
+  if(NOT BUILD_CAFFE2 AND NOT INTERN_BUILD_MOBILE AND NOT INTERN_DISABLE_ONNX)
     list(APPEND Caffe2_CPU_SRCS
       utils/proto_wrap.cc
     )
